fix windows compat and remove no-color warning
We're going to assume that anyone interested in this program will have
some level of color support, so we default to 16 color support if curses
can't detect another.

This also fixes a spurious "no color support" warning on Windows when it
should work (term color support works different on windows). So this
also fixes that issue.